requirements.txt
fastapi
uvicorn
sqlmodel
requests
pandas
torch
transformers
scikit-learn
matplotlib
statsmodels
python-multipart
pyjwt[crypto]
cryptography
python-dotenv
feedparser
openai>=1.0
beautifulsoup4
lxml
